Implement GcBiomeData
