+++ MBINCompiler/Models/Structs/GcExternalObjectListOptions.cs
@@ -1,0 +1,15 @@
+﻿namespace MBINCompiler.Models.Structs
+{
+    public class GcExternalObjectListOptions : NMSTemplate
+    {
+        [NMS(Size = 0x10)]
+        public string Name;
+        public float Probability;
+        public GcTerrainTileType TileType;
+        [NMS(Size = 4)]
+        public NMSString0x80[] Options;
+        public bool AllowLimiting;
+        [NMS(Size = 7, Ignore = true)]
+        public byte[] Padding;
+    }
+}
