+++ MBINCompiler/Models/Structs/GcHeavyAirColourData.cs
@@ -1,0 +1,8 @@
+﻿namespace MBINCompiler.Models.Structs
+{
+    public class GcHeavyAirColourData : NMSTemplate
+    {
+        public Colour Colour1;
+        public Colour Colour2;
+    }
+}
