+++ MBINCompiler/Models/Structs/GcMiningSubstanceData.cs
@@ -1,0 +1,9 @@
+﻿namespace MBINCompiler.Models.Structs
+{
+    public class GcMiningSubstanceData : NMSTemplate
+    {
+        public GcRealitySubstanceCategory Category;
+        public bool UseRarity;
+        public GcRarity Rarity;
+    }
+}
