+++ MBINCompiler/Models/Structs/GcPlanetHeavyAirData.cs
@@ -1,0 +1,11 @@
+﻿namespace MBINCompiler.Models.Structs
+{
+    public class GcPlanetHeavyAirData : NMSTemplate
+    {
+        [NMS(Size = 0x80)]
+        public string Filename;
+        [NMS(Size = 5)]
+        public GcHeavyAirColourData[] Colours;
+    }
+
+}
