+++ MBINCompiler/Models/Structs/GcTileTypeOptions.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace MBINCompiler.Models.Structs
+{
+    public class GcTileTypeOptions : NMSTemplate
+    {
+        public List<TkPaletteTexture> Options;
+    }
+}
