+++ MBINCompiler/Models/Structs/NMSBool.cs
@@ -1,0 +1,9 @@
+﻿using System.Runtime.InteropServices;
+
+namespace MBINCompiler.Models.Structs
+{
+    public class NMSBool : NMSTemplate
+    {
+        public bool Value;
+    }
+}
