+++ MBINCompilerTests/UnitTests.cs
@@ -11875,6 +11875,15 @@ namespace MBINCompilerTests
 
             // TkVoxelGeneratorSettingsArray
             "METADATA/SIMULATION/SOLARSYSTEM/VOXELGENERATORSETTINGS.MBIN",
+
+            // GcBiomeData
+            "METADATA/SIMULATION/SOLARSYSTEM/BIOMES/BARREN/BARRENBIOME.MBIN",
+            "METADATA/SIMULATION/SOLARSYSTEM/BIOMES/DEAD/DEADBIOME.MBIN",
+            "METADATA/SIMULATION/SOLARSYSTEM/BIOMES/FROZEN/FROZENBIOME.MBIN",
+            "METADATA/SIMULATION/SOLARSYSTEM/BIOMES/LUSH/LUSHBIOME.MBIN",
+            "METADATA/SIMULATION/SOLARSYSTEM/BIOMES/RADIOACTIVE/RADIOACTIVEBIOME.MBIN",
+            "METADATA/SIMULATION/SOLARSYSTEM/BIOMES/SCORCHED/SCORCHEDBIOME.MBIN",
+            "METADATA/SIMULATION/SOLARSYSTEM/BIOMES/TOXIC/TOXICBIOME.MBIN",
         };
 
         private void decompileFile(string filePath, string output = "")
